requirements.txt
Django==4.1.7
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
openai==0.27.2
requests==2.28.2
django-allauth==0.52.0
Pillow==9.5.0
PyMuPDF==1.22.1
pywin32==306; platform_system == "Windows"
python-decouple==3.8
